+++ packages/ripple-address-codec/HISTORY.md
@@ -1,5 +1,14 @@
 # ripple-address-codec Release History
 
+## 4.1.0 (UNRELEASED)
+
+* Throwable 'unexpected_payload_length' error: The message has been expanded with ' Ensure that the bytes are a Buffer.'
+
+### New Features
+
+* `encodeAccountPublic` - Encode a public key, as for payment channels
+* `decodeAccountPublic` - Decode a public key, as for payment channels
+
 ## 4.0.0 (2019-10-08)
 
 ### Breaking Changes
@@ -14,4 +23,3 @@
 * `xAddressToClassicAddress` - Decode an X-address to account ID, tag, and network ID
 * `decodeXAddress` - Convert X-address to classic address, tag, and network ID
 * `isValidXAddress` - Check whether an X-address (X...) is valid
-